% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidyverse_deps}
\alias{tidyverse_deps}
\title{List all tidyverse dependencies}
\usage{
tidyverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all tidyverse dependencies
}
