% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{tidyverse_logo}
\alias{tidyverse_logo}
\title{The tidyverse logo, using ASCII or Unicode characters}
\usage{
tidyverse_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[cli:ansi_strip]{cli::ansi_strip()}} to get rid of the colors.
}
\examples{
tidyverse_logo()
}
