% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\docType{package}
\name{tidyverse-package}
\alias{tidyverse}
\alias{tidyverse-package}
\title{tidyverse: Easily Install and Load 'Tidyverse' Packages}
\description{
The 'tidyverse' is a set of packages that work in harmony
because they share common data representations and 'API' design. This
package is designed to make it easy to install and load multiple
'tidyverse' packages in a single step. Learn more about the 'tidyverse'
at <https://github.com/hadley/tidyverse>.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://tidyverse.org}
  \item \url{https://github.com/tidyverse/tidyverse}
  \item Report bugs at \url{https://github.com/tidyverse/tidyverse/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
