% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidyverse_deps}
\alias{tidyverse_deps}
\title{List all tidyverse dependencies}
\usage{
tidyverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}
}
\description{
List all tidyverse dependencies
}
