% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidyverse_update}
\alias{tidyverse_update}
\title{Update tidyverse packages}
\usage{
tidyverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
This will check to see if all tidyverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidyverse_update()
}
}
